Add unit test for #5
