+++ tests/YmlTransform.Tests/Data/Unversioned.yml
@@ -1,0 +1,25 @@
+﻿---
+ID: "ab5b233c-111c-4fa8-affe-b0c7315fedf4"
+Parent: "756fff26-90f2-4889-bace-da656e66b9c1"
+Template: "013bc834-c0f6-4377-924e-0de185abfbe7"
+Path: /sitecore/content/Home/Unversioned
+DB: master
+SharedFields:
+- ID: "ba3f86a2-4a1c-4d78-b63d-91c2779c1b5e"
+  Hint: __Sortorder
+  Value: 375
+Languages:
+- Language: en
+  Fields:
+  - ID: "11c0c7c6-673f-432a-b2fa-acfead30b6dd"
+    Hint: Field1
+    Value: "Value1"
+  - ID: "814e27bf-7ae2-433c-affa-aafc0797e0a8"
+    Hint: Field2
+    Value: "Value2"
+Versions:
+  - Version: 1
+    Fields:
+    - ID: "25bed78c-4957-4165-998a-ca1b52f67497"
+      Hint: __Created
+      Value: 20171016T073039Z
