+++ tests/YmlTransform.Tests/packages.config
@@ -2,6 +2,7 @@
 <packages>
   <package id="System.Reflection" version="4.3.0" targetFramework="net47" />
   <package id="System.Runtime" version="4.3.0" targetFramework="net47" />
+  <package id="System.Runtime.Extensions" version="4.3.0" targetFramework="net47" />
   <package id="xunit" version="2.3.0" targetFramework="net47" />
   <package id="xunit.abstractions" version="2.0.1" targetFramework="net47" />
   <package id="xunit.analyzers" version="0.7.0" targetFramework="net47" />
